backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
joblib==1.4.2
